Change get/setFlag to use Int32 for MPV_FORMAT_FLAG (#5824)
This commit will:
- Change MPVController.getFlag to use Int32 instead of Int64 when
  calling mpv_get_property
- Change MPVController.setFlag to use Int32 instead of Int when calling
  mpv_set_property

Using 64 bit integers is working, but is not technically correct. A
MPV_FORMAT_FLAG is a 32 bit integer. MPVController correctly uses Int32
for other libmpv client APIs that use 32 bit integers. These changes
align these two methods with that convention.

+++ iina/MPVController.swift
@@ -254,7 +254,7 @@ final class MPVController: NSObject {
 
   func setFlag(_ name: String, _ flag: Bool, level: Logger.Level = .debug) {
     log.log("Set property: \(name)=\(flag.yesno)", level: level)
-    var data: Int = flag ? 1 : 0
+    var data: Int32 = flag ? 1 : 0
     guard mpv != nil else { log.warn("Aborting setProperty: mpv is nil"); return }
     let code = mpv_set_property(mpv, name, MPV_FORMAT_FLAG, &data)
     if code < 0 {
@@ -311,7 +311,7 @@ final class MPVController: NSObject {
   }
 
   func getFlag(_ name: String) -> Bool {
-    var data = Int64()
+    var data = Int32()
     mpv_get_property(mpv, name, MPV_FORMAT_FLAG, &data)
     return data > 0
   }
